lesson test style fixed , component managed
